리팩토링 : 커스텀훅 & 컴포넌트 분리
